+++ NEWS.md
@@ -1,5 +1,7 @@
 # orbis 0.1.0.9000 Development version
 
+- Removed `get_wc_osf_id()` function as it is no longer needed.
+
 # orbis 0.1.0
 
 - First release. 🎉